refactor: プレゼンテーション層の全体的な設定は、直下に移動。
